Code: Path fixes for [Le/Sa]ss compilers. * Passes user setting to Less compiler. * Sass compiler does not provide the setting, so I have moved   `PostProcessResult()` to `SassCompiler` derived class.
Next we need to stitch the decoded source-maps into normalizer to get
the correct paths in import scenarios.
